refactor(utils): categorize functions into separate files
