+++ packages/docusaurus-utils/src/__tests__/urlUtils.test.ts
@@ -5,7 +5,17 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-import {normalizeUrl, getEditUrl} from '../urlUtils';
+import {
+  normalizeUrl,
+  getEditUrl,
+  fileToPath,
+  isValidPathname,
+  addTrailingSlash,
+  addLeadingSlash,
+  removeTrailingSlash,
+  resolvePathname,
+  encodePath,
+} from '../urlUtils';
 
 describe('normalizeUrl', () => {
   test('should normalize urls correctly', () => {
@@ -149,4 +159,83 @@ describe('getEditUrl', () => {
   test('returns undefined for undefined', () => {
     expect(getEditUrl('foo/bar.md')).toBeUndefined();
   });
+});
+
+test('fileToPath', () => {
+  const asserts: Record<string, string> = {
+    'index.md': '/',
+    'hello/index.md': '/hello/',
+    'foo.md': '/foo',
+    'foo/bar.md': '/foo/bar',
+    'index.js': '/',
+    'hello/index.js': '/hello/',
+    'foo.js': '/foo',
+    'foo/bar.js': '/foo/bar',
+  };
+  Object.keys(asserts).forEach((file) => {
+    expect(fileToPath(file)).toBe(asserts[file]);
+  });
+});
+
+test('isValidPathname', () => {
+  expect(isValidPathname('/')).toBe(true);
+  expect(isValidPathname('/hey')).toBe(true);
+  expect(isValidPathname('/hey/ho')).toBe(true);
+  expect(isValidPathname('/hey/ho/')).toBe(true);
+  expect(isValidPathname('/hey/h%C3%B4/')).toBe(true);
+  expect(isValidPathname('/hey///ho///')).toBe(true); // Unexpected but valid
+  expect(isValidPathname('/hey/héllô you')).toBe(true);
+
+  expect(isValidPathname('')).toBe(false);
+  expect(isValidPathname('hey')).toBe(false);
+  expect(isValidPathname('/hey?qs=ho')).toBe(false);
+  expect(isValidPathname('https://fb.com/hey')).toBe(false);
+  expect(isValidPathname('//hey')).toBe(false);
+  expect(isValidPathname('////')).toBe(false);
+});
+
+describe('addTrailingSlash', () => {
+  test('should no-op', () => {
+    expect(addTrailingSlash('/abcd/')).toEqual('/abcd/');
+  });
+  test('should add /', () => {
+    expect(addTrailingSlash('/abcd')).toEqual('/abcd/');
+  });
+});
+
+describe('addLeadingSlash', () => {
+  test('should no-op', () => {
+    expect(addLeadingSlash('/abc')).toEqual('/abc');
+  });
+  test('should add /', () => {
+    expect(addLeadingSlash('abc')).toEqual('/abc');
+  });
+});
+
+describe('removeTrailingSlash', () => {
+  test('should no-op', () => {
+    expect(removeTrailingSlash('/abcd')).toEqual('/abcd');
+  });
+  test('should remove /', () => {
+    expect(removeTrailingSlash('/abcd/')).toEqual('/abcd');
+  });
+});
+
+test('resolvePathname', () => {
+  // These tests are directly copied from https://github.com/mjackson/resolve-pathname/blob/master/modules/__tests__/resolvePathname-test.js
+  // Maybe we want to wrap that logic in the future?
+  expect(resolvePathname('c')).toEqual('c');
+  expect(resolvePathname('c', 'a/b')).toEqual('a/c');
+  expect(resolvePathname('/c', '/a/b')).toEqual('/c');
+  expect(resolvePathname('', '/a/b')).toEqual('/a/b');
+  expect(resolvePathname('../c', '/a/b')).toEqual('/c');
+  expect(resolvePathname('c', '/a/b')).toEqual('/a/c');
+  expect(resolvePathname('c', '/a/')).toEqual('/a/c');
+  expect(resolvePathname('..', '/a/b')).toEqual('/');
+});
+
+test('encodePath', () => {
+  expect(encodePath('a/foo/')).toEqual('a/foo/');
+  expect(encodePath('a/<foo>/')).toEqual('a/%3Cfoo%3E/');
+  expect(encodePath('a/你好/')).toEqual('a/%E4%BD%A0%E5%A5%BD/');
 });
